fix compiler warnings in garnet_thief, rainbow_treasure_hunter, long_night, crowd
- garnet_thief, rainbow_treasure_hunter: .is_formal_{false} -> .is_formal_ = 0
- long_night/player.h: unqualified move() -> std::move()
- long_night/mygame.cc: add NONE/RANDOM cases to silence -Wswitch
- crowd/problems.h: =- 0.5 -> = -0.5 to silence unary-operator warning

+++ match_process/child_config_session.cc
@@ -198,7 +198,7 @@ bool ChildConfigSession::HandleInitOptions(const lgtbot::ipc::InitOptionsReq& re
     }
 
     lgtbot::game::MutableGenericOptions generic_options{};
-    generic_options.is_formal_ = static_cast<uint8_t>(default_is_formal_);
+    generic_options.is_formal_ = default_is_formal_;
     const auto result = module_.init_options_(req.args().c_str(), opts.get(), &generic_options);
 
     lgtbot::ipc::ConfigResponse resp;
